+++ src/index.ts
@@ -1,5 +1,11 @@
 import type { App, InjectionKey } from 'vue'
 import type { OAuthClientOptions } from './OAuthClient'
+import {
+    ClientSecretBasic,
+    ClientSecretPost,
+    PrivateKeyJwt,
+    TlsClientAuth,
+} from 'oauth4webapi'
 import { getCurrentInstance, inject } from 'vue'
 import { OAuthClient } from './OAuthClient'
 
@@ -7,6 +13,12 @@ export { OAuthClient }
 export { LocalStorage } from './LocalStorage'
 export { SessionStorage } from './SessionStorage'
 export { Storage } from './Storage'
+export {
+    ClientSecretBasic,
+    ClientSecretPost,
+    PrivateKeyJwt,
+    TlsClientAuth,
+}
 
 export const authClientInjectionKey = Symbol('') as InjectionKey<OAuthClient>
 
@@ -77,3 +89,5 @@ export function useOAuthClient(options?: OAuthClientOptions) {
     }
     return client
 }
+
+export {}