+++ test/oAuthClient.test.ts
@@ -5,13 +5,13 @@ import { OAuthClient } from '../src/OAuthClient'
 const fetchMock = createFetchMock(vi)
 
 const response = JSON.stringify({
-    token_endpoint: 'http://dummy.com/oauth2/v2.0/token',
+    token_endpoint: 'https://dummy.com/oauth2/v2.0/token',
     token_endpoint_auth_methods_supported: [
         'client_secret_post',
         'private_key_jwt',
         'client_secret_basic',
     ],
-    jwks_uri: 'http://dummy.com/discovery/v2.0/keys',
+    jwks_uri: 'https://dummy.com/discovery/v2.0/keys',
     response_modes_supported: ['query', 'fragment', 'form_post'],
     subject_types_supported: ['pairwise'],
     id_token_signing_alg_values_supported: ['RS256'],
@@ -22,14 +22,14 @@ const response = JSON.stringify({
         'id_token token',
     ],
     scopes_supported: ['openid', 'profile', 'email', 'offline_access'],
-    issuer: 'http://dummy.com',
+    issuer: 'https://dummy.com',
     request_uri_parameter_supported: false,
     userinfo_endpoint: 'https://graph.microsoft.com/oidc/userinfo',
-    authorization_endpoint: 'http://dummy.com/oauth2/v2.0/authorize',
-    device_authorization_endpoint: 'http://dummy.com/oauth2/v2.0/devicecode',
+    authorization_endpoint: 'https://dummy.com/oauth2/v2.0/authorize',
+    device_authorization_endpoint: 'https://dummy.com/oauth2/v2.0/devicecode',
     http_logout_supported: true,
     frontchannel_logout_supported: true,
-    end_session_endpoint: 'http://dummy.com/oauth2/v2.0/logout',
+    end_session_endpoint: 'https://dummy.com/oauth2/v2.0/logout',
     claims_supported: [
         'sub',
         'iss',
@@ -51,7 +51,7 @@ const response = JSON.stringify({
         'c_hash',
         'email',
     ],
-    kerberos_endpoint: 'http://dummy.com/kerberos',
+    kerberos_endpoint: 'https://dummy.com/kerberos',
     tenant_region_scope: 'EU',
     cloud_instance_name: 'microsoftonline.com',
     cloud_graph_host_name: 'graph.windows.net',
@@ -67,7 +67,7 @@ describe('oAuthClient', () => {
     it('should create a new OAuthClient instance', () => {
         const client = new OAuthClient({
             clientId: '864b865f-3025-4c48-b4ed-c16676a5b676',
-            url: 'http://dummy.com',
+            url: 'https://dummy.com',
         })
         expect(client).toBeInstanceOf(OAuthClient)
     })
@@ -76,12 +76,12 @@ describe('oAuthClient', () => {
         fetchMock.mockResponse(response, {
             status: 200,
             statusText: 'ok',
-            url: 'http://dummy.com/.well-known/openid-configuration',
+            url: 'https://dummy.com/.well-known/openid-configuration',
         })
         const client = new OAuthClient({
             clientId: 'test',
             scopes: ['test'],
-            url: 'http://dummy.com',
+            url: 'https://dummy.com',
         })
         expect(client).toBeInstanceOf(OAuthClient)
         await client.initialize()
@@ -99,19 +99,19 @@ describe('oAuthClient', () => {
         fetchMock.mockResponse(response, {
             status: 200,
             statusText: 'ok',
-            url: 'http://dummy.com/.well-known/openid-configuration',
+            url: 'https://dummy.com/.well-known/openid-configuration',
         })
         const client = new OAuthClient({
             clientId: 'test',
             scopes: ['test'],
-            url: 'http://dummy.com',
+            url: 'https://dummy.com',
         })
         expect(client).toBeInstanceOf(OAuthClient)
         await client.initialize()
         expect(client.initialized).toBe(true)
         await client.authorize()
         const urlToTest = mockResponse.mock.calls[0][0]
-        expect(urlToTest).toContain('http://dummy.com/oauth2/v2.0/authorize')
+        expect(urlToTest).toContain('https://dummy.com/oauth2/v2.0/authorize')
         expect(urlToTest).toContain('scope=test')
         expect(urlToTest).toContain('client_id=test')
     })